Refactor our global styles into their own component
